utility function update

+++ congestion/vivace.go
@@ -10,8 +10,8 @@ import (
 type VivaceInterval struct {
 	start	time.Time
 	duration time.Duration
-	rate uint32
-	sendingRate uint32
+	rate float64
+	sendingRate float64
 	sentPackets uint32
 	lostPackets uint32
 	ackedPackets uint32
@@ -23,7 +23,7 @@ func NewVivaceInterval (duration time.Duration,currentTime time.Time) VivaceInte
 		duration: duration,
 		rate: 3,
 		sendingRate: 0,
-		sentPackets: 0, 
+		sentPackets: 0,
 		lostPackets: 0,
 		ackedPackets: 0,
 		}
